requirements.txt
crewai
crewai_tools
openai
streamlit
langsmith
langchain
langchain-community
chromadb
PyPDF2
python-dotenv
openai
pypdf
tiktoken
langchain-openai 
langchain-chroma
fitz
PyMuPDF
pymupdf
streamlit
flask
Flask-CORS
langchain_chroma
langchain_openai
langchain-groq
langchain_core
langtrace-python-sdk
azure-storage-blob